requirements.txt
matplotlib==3.4.3
numpy==1.20.3
Pillow==9.1.0
requests==2.26.0
plotly==5.7.0
